OLD
```bash
Benchmark_Ctx_Attachment-4       3026473               392 ns/op             208 B/op          3 allocs/op
Benchmark_Ctx_Attachment-4       3047916               392 ns/op             208 B/op          3 allocs/op
Benchmark_Ctx_Attachment-4       2982228               392 ns/op             208 B/op          3 allocs/op
Benchmark_Ctx_Attachment-4       3003832               393 ns/op             208 B/op          3 allocs/op
```

NEW
```bash
Benchmark_Ctx_Attachment-4       3225854               367 ns/op             208 B/op          2 allocs/op
Benchmark_Ctx_Attachment-4       3215389               368 ns/op             208 B/op          2 allocs/op
Benchmark_Ctx_Attachment-4       3242175               369 ns/op             208 B/op          2 allocs/op
Benchmark_Ctx_Attachment-4       3146233               367 ns/op             208 B/op          2 allocs/op
```